+++ lib/src/configuration/configuration_ipv6_dhcp.dart
@@ -1,5 +1,5 @@
 import 'package:cisco_ios_dsl/src/device.dart';
-import 'package:cisco_ios_dsl/src/ipv4_pool/ipv4_pool_scope.dart';
+import 'package:cisco_ios_dsl/src/ipv6_pool/ipv6_pool_scope.dart';
 import 'package:cisco_ios_dsl/src/util/scope.dart';
 
 class ConfigurationIpv6Dhcp {
@@ -7,6 +7,6 @@ class ConfigurationIpv6Dhcp {
   final Device device;
   ConfigurationIpv6Dhcp(this.device);
 
-  void pool(String name, void Function(IPv4PoolScope x) body) => device
+  void pool(String name, void Function(IPv6PoolScope x) body) => device
     ..useScope(Scope.dhcpv6(), 'ipv6 dhcp pool $name', body);
 }